Firefox OS: open external links in browser app
Using window.open() opened a new popup window within the Wikipedia app,
which seemed a little weird.

Using MozActivity class, we can launch the main browser application
with the given URL, which is what we think we want here. I think. :)

Note the hack to remove 'proxy.php?url=' -- this should not be happening,
there seems to be some aggressive addition of the proxy URL somewhere
where it shouldn't be. Remove this once fixed.